+++ trunk/include/runinfo_def.h
@@ -55,8 +55,8 @@ struct runinfo_struct {
   // output global values
   int do_global_output;  // set if global, multiwavelength output is desired
   int out_sed_lum_offset; // offset for vector to all for ERE/Dust direct/scattered to be saved as well as stellar direct/scattered
-  std::vector< std::vector<float> > out_sed_lum;  // allow for direct/scattered lum
-  std::vector< std::vector<float> > out_sed_lum_unc;  // allow for direct/scattered lum uncertainties
+  std::vector< std::vector<double> > out_sed_lum;  // allow for direct/scattered lum
+  std::vector< std::vector<double> > out_sed_lum_unc;  // allow for direct/scattered lum uncertainties
 /*   std::vector< std::vector<float> > out_emitted_lum;  // allow to be by grain/emission type */
 };
